Rename useSectionRatio as useScrolledRatio
