+++ docs/src/ScrollTracking/Section.mdx
@@ -100,18 +100,18 @@ What does `sectionInfo` provides?
 
 If you don't want to use the render props to get `sectionInfo`, you can also use the hooks to get the same information by yourself:
 
-### `useSectionRatio` hook
+### `useScrolledRatio` hook
 
 Import the hooks:
 ```jsx
-import { useIntersectionObservable, useSectionRatio } from '@react-scrolly/core';
+import { useIntersectionObservable, useScrolledRatio } from '@react-scrolly/core';
 ```
 
 Use them in your functional components:
 ```jsx
 const sectionRef = useRef(null);
 const intersectObsr$ = useIntersectionObservable(sectionRef, trackingId);
-const sectionInfo = useSectionRatio(sectionRef, intersectObsr$, trackingId);
+const sectionInfo = useScrolledRatio(sectionRef, intersectObsr$, trackingId);
 
 return (
   <div
@@ -137,7 +137,7 @@ By adding `trackingId` to sections, `<PageProvider>` select the section closest 
 
 or
 ```jsx
-const sectionInfo = useSectionRatio(sectionRef, intersectObsr$, 'TRACKING_ID');
+const sectionInfo = useScrolledRatio(sectionRef, intersectObsr$, 'TRACKING_ID');
 ```
 
 ## Complete Example
