+++ packages/plot/src/components/Plot.tsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import {
   // hooks
   useIntersectionObservable,
-  useSectionRatio,
+  useScrolledRatio,
   // types
   SectionProps,
  } from '@react-scrolly/core';
@@ -18,7 +18,7 @@ export const Plot = ({
 }: SectionProps) => {
   const sectionRef = useRef<HTMLDivElement>(null);
   const intersectObsr$ = useIntersectionObservable(sectionRef, trackingId);
-  const sectionInfo = useSectionRatio(sectionRef, intersectObsr$, trackingId);
+  const sectionInfo = useScrolledRatio(sectionRef, intersectObsr$, trackingId);
   const activeSection = useActiveSectionInfo(intersectObsr$);
 
   return (
